Break Overrides components into individual files
